Move utility functions to `utils.js`, document, and prettify
